Convert RelocationCachingSpec
